+++ src/CommandInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace nyansapow;
+
+/**
+ * Description of CommandInterface
+ *
+ * @author ekow
+ */
+interface CommandInterface {
+    public function execute($options);
+}